+++ CHANGELOG.md
@@ -1,3 +1,18 @@
 ## [0.0.1]
 
 * First version of the package!
+
+## [0.0.2]
+
+* Adding examples
+* The code is now formatted with DartFM 
+
+## [0.0.3]
+
+* Renaming lib/date_range.dart to lib/date_range_common.dart to match the package name
+
+
+## [0.1.0]
+
+* Fixing issue with the get_duration method (start <-> end)
+* Adding documentation